change sign in button to separate from account dropdown, change router in Logout, minor styling change
